fix(rn): isolate live inference worklet boundary
