+++ docs/public/guides/public-api.md
@@ -203,10 +203,11 @@ renderer is introduced.
 The remaining mobile work is integration, not a second session abstraction:
 migrate the saved-video compatibility factory onto the generic
 source/processor/renderer contracts and add an Android saved-video source before
-claiming cross-platform file support. The experimental `./react` entrypoint now
-owns the live VisionCamera worklet through `useReactNativeLiveInference()`;
-applications provide model runners and serializable configuration, not frame
-callbacks.
+claiming cross-platform file support. The experimental
+`./react/live-inference` entrypoint owns the live VisionCamera worklet through
+`useReactNativeLiveInference()`; applications provide model runners and
+serializable configuration, not frame callbacks. This keeps the generic
+`./react` entry usable without the optional Worklets peer.
 
 ## Compatibility Posture
