+++ packages/react-native/src/react/index.ts
@@ -16,24 +16,7 @@ export {
 } from "./use-media-session";
 export { useReactNativeLiveCameraBinding } from "./use-live-camera-binding";
 export { useReactNativeClassMaskEffects } from "./use-class-mask-effects";
-export {
-  useReactNativeLiveInference,
-  type ReactNativeLiveClassEffect,
-  type ReactNativeLiveClassEffects,
-  type ReactNativeLiveInferenceBinding,
-  type ReactNativeLiveInferenceDetection,
-  type ReactNativeLiveInferenceError,
-  type ReactNativeLiveInferenceExtensionOptions,
-  type ReactNativeLiveInferenceInteractionRequest,
-  type ReactNativeLiveInferenceInteractionResult,
-  type ReactNativeLiveInferenceMode,
-  type ReactNativeLiveInferenceReadout,
-  type UseReactNativeLiveInferenceOptions,
-} from "./use-live-inference";
-export {
-  scheduleReactNativeOnJs,
-  useReactNativeSharedValue,
-} from "./worklet-bridge";
+export { useReactNativeSharedValue } from "./worklet-bridge";
 export {
   createReactNativeLiveStageOverlays,
   createReactNativeLiveDetectionStageOverlays,
